+++ superset-frontend/src/types/bootstrapTypes.ts
@@ -8,6 +8,7 @@ import {
 } from '@superset-ui/core';
 import { isPlainObject } from 'lodash';
 import { FlashMessage } from '../components/FlashProvider';
+import { Languages } from '../views/components/LanguagePicker';
 
 /**
  * Licensed to the Apache Software Foundation (ASF) under one
@@ -82,6 +83,59 @@ export type ChartResponse = {
   result: ChartData[];
 };
 
+export interface BrandProps {
+  path: string;
+  icon: string;
+  alt: string;
+  tooltip: string;
+  text: string;
+}
+
+export interface NavBarProps {
+  show_watermark: boolean;
+  bug_report_url?: string;
+  version_string?: string;
+  version_sha?: string;
+  build_number?: string;
+  documentation_url?: string;
+  languages: Languages;
+  show_language_picker: boolean;
+  user_is_anonymous: boolean;
+  user_info_url: string;
+  user_login_url: string;
+  user_logout_url: string;
+  user_profile_url: string | null;
+  locale: string;
+}
+
+export interface MenuObjectChildProps {
+  label: string;
+  name?: string;
+  icon?: string;
+  index?: number;
+  url?: string;
+  isFrontendRoute?: boolean;
+  perm?: string | boolean;
+  view?: string;
+  disable?: boolean;
+}
+
+export interface MenuObjectProps extends MenuObjectChildProps {
+  childs?: (MenuObjectChildProps | string)[];
+  isHeader?: boolean;
+}
+
+export interface MenuData {
+  menu: MenuObjectProps[];
+  brand: BrandProps;
+  navbar_right: NavBarProps;
+  settings: MenuObjectProps[];
+  environment_tag: {
+    text: string;
+    color: string;
+  };
+}
+
 export interface CommonBootstrapData {
   flash_messages: FlashMessage[];
   conf: JsonObject;
@@ -91,7 +145,7 @@ export interface CommonBootstrapData {
   extra_categorical_color_schemes: ColorSchemeConfig[];
   extra_sequential_color_schemes: SequentialSchemeConfig[];
   theme_overrides: JsonObject;
-  menu_data: JsonObject;
+  menu_data: MenuData;
 }
 
 export interface BootstrapData {
