+++ superset-frontend/src/utils/getBootstrapData.ts
@@ -17,19 +17,11 @@
  * under the License.
  */
 
-import { useEffect, useState } from 'react';
-
 import { BootstrapData } from 'src/types/bootstrapTypes';
 
-export function useBootstrapData(): BootstrapData {
-  const [bootstrapData, setBootstrapData] = useState<BootstrapData>({});
+export default function getBootstrapData(): BootstrapData {
+  const appContainer = document.getElementById('app');
+  const dataBootstrap = appContainer?.getAttribute('data-bootstrap') || '{}';
 
-  useEffect(() => {
-    const appContainer = document.getElementById('app');
-    const dataBootstrap = appContainer?.getAttribute('data-bootstrap') || '{}';
-
-    setBootstrapData(JSON.parse(dataBootstrap));
-  }, []);
-
-  return bootstrapData;
+  return JSON.parse(dataBootstrap);
 }
