+++ superset-frontend/src/views/components/Menu.tsx
@@ -29,63 +29,16 @@ import { GenericLink } from 'src/components/GenericLink/GenericLink';
 import Icons from 'src/components/Icons';
 import { useUiConfig } from 'src/components/UiConfigContext';
 import { URL_PARAMS } from 'src/constants';
+import {
+  MenuObjectChildProps,
+  MenuObjectProps,
+  MenuData,
+} from 'src/types/bootstrapTypes';
 import RightMenu from './RightMenu';
-import { Languages } from './LanguagePicker';
 
-interface BrandProps {
-  path: string;
-  icon: string;
-  alt: string;
-  tooltip: string;
-  text: string;
-}
-
-export interface NavBarProps {
-  show_watermark: boolean;
-  bug_report_url?: string;
-  version_string?: string;
-  version_sha?: string;
-  build_number?: string;
-  documentation_url?: string;
-  languages: Languages;
-  show_language_picker: boolean;
-  user_is_anonymous: boolean;
-  user_info_url: string;
-  user_login_url: string;
-  user_logout_url: string;
-  user_profile_url: string | null;
-  locale: string;
-}
-
-export interface MenuProps {
-  data: {
-    menu: MenuObjectProps[];
-    brand: BrandProps;
-    navbar_right: NavBarProps;
-    settings: MenuObjectProps[];
-    environment_tag: {
-      text: string;
-      color: string;
-    };
-  };
+interface MenuProps {
+  data: MenuData;
   isFrontendRoute?: (path?: string) => boolean;
-}
-
-export interface MenuObjectChildProps {
-  label: string;
-  name?: string;
-  icon?: string;
-  index?: number;
-  url?: string;
-  isFrontendRoute?: boolean;
-  perm?: string | boolean;
-  view?: string;
-  disable?: boolean;
-}
-
-export interface MenuObjectProps extends MenuObjectChildProps {
-  childs?: (MenuObjectChildProps | string)[];
-  isHeader?: boolean;
 }
 
 const StyledHeader = styled.header`
